requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==19.3.0
Automat==0.8.0
beautifulsoup4==4.12.3
blinker==1.4
Brotli==1.1.0
bs4==0.0.2
cachetools==5.3.2
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.3.2
Click==7.0
cloud-init==23.3.3
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
discord==2.3.2
discord.py==2.3.2
distro==1.4.0
distro-info==0.23+ubuntu1.1
entrypoints==0.3
frozenlist==1.4.1
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
multidict==6.0.4
mutagen==1.47.0
netifaces==0.10.4
oauthlib==3.1.0
pbr==6.0.0
pexpect==4.6.0
protobuf==4.25.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycryptodomex==3.20.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-debian==0.1.36+ubuntu1.1
PyYAML==5.3.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.5.6
soupsieve==2.5
ssh-import-id==5.10
systemd-python==234
testresources==2.0.1
Twisted==18.9.0
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.1.0
wadllib==1.3.3
websockets==12.0
yarl==1.9.4
yt-dlp==2023.12.30
zipp==1.0.0
zope.interface==4.7.1
